requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.25.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
decorator==5.2.1
dm-tree==0.1.8
dp-accounting==0.4.3
flatbuffers==25.2.10
gast==0.6.0
gmpy2==2.2.1
grpcio==1.73.0
h5py==3.14.0
httplib2==0.22.0
idna==3.10
isodate==0.7.2
joblib==1.5.1
keras==2.14.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
mpyc==0.10
namex==0.1.0
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==22.0
pandas==2.3.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
tabulate==0.9.0
tensorflow==2.14.1
tensorflow-probability==0.22.1
tensorflow_privacy==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
aiofiles==24.1.0
Werkzeug==3.1.3
wrapt==1.14.1
wandb==0.22.0